Database_connecting code
It is the PHP code to connect the database to the html code named "LoginForm".  Here the database will be connected to the form and check hte values of name and password from the table from Database.  I have written the descriptions beside the code what i have done in which line.